+++ src/main/java/com/codessquad/qna/repository/QuestionRepository.java
@@ -1,5 +1,6 @@
-package com.codessquad.qna.questions;
+package com.codessquad.qna.repository;
 
+import com.codessquad.qna.entity.Question;
 import org.springframework.data.repository.CrudRepository;
 
 public interface QuestionRepository extends CrudRepository<Question, Long> {
