+++ src/main/java/com/codessquad/qna/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.codessquad.qna.users;
+package com.codessquad.qna.repository;
 
+import com.codessquad.qna.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
